saved work on client generator
