+++ src/Client/Generators/ITypeLookup.cs
@@ -7,7 +7,7 @@ namespace StrawberryShake.Generators
     {
         string GetTypeName(FieldNode field, IType fieldType, bool readOnly);
 
-        ITypeInfo GetTypeInfo(FieldNode field, IType fieldType, bool readOnly);
+        ITypeInfo GetTypeInfo(IType fieldType, bool readOnly);
 
         string GetTypeName(IType fieldType, string typeName, bool readOnly);
     }
